update title if flags change
