+++ js/models/folder.js
@@ -5,7 +5,7 @@
  * later. See the COPYING file.
  *
  * @author Christoph Wurst <christoph@winzerhof-wurst.at>
- * @copyright Christoph Wurst 2015, 2016
+ * @copyright Christoph Wurst 2015
  */
 
 define(function(require) {
